Use TryFrom for throughput-to-headers conversion
Replace standalone throughput_headers module with a TryFrom impl
on ThroughputProperties, keeping conversion logic co-located with
the type it converts. This is the idiomatic Rust approach for
fallible type conversions.
