+++ sdk/cosmos/azure_data_cosmos/src/lib.rs
@@ -58,4 +58,3 @@ mod retry_policies;
 mod routing;
 mod routing_strategy;
 mod serde;
-mod throughput_headers;
