+++ sdk/cosmos/azure_data_cosmos/src/models/throughput_properties.rs
@@ -138,3 +138,61 @@ pub(crate) struct AutoscaleAutoUpgradePolicy {
 pub(crate) struct AutoscaleThroughputPolicy {
     pub increment_percent: usize,
 }
+
+impl TryFrom<&ThroughputProperties> for azure_data_cosmos_driver::models::CosmosRequestHeaders {
+    type Error = azure_core::Error;
+
+    fn try_from(throughput: &ThroughputProperties) -> Result<Self, Self::Error> {
+        let mut headers = Self::new();
+
+        match (
+            throughput.throughput(),
+            throughput.autoscale_maximum().is_some(),
+        ) {
+            (Some(t), _) => {
+                headers.offer_throughput = Some(t);
+            }
+            (_, true) => {
+                let autopilot_json =
+                    serde_json::to_string(&throughput.offer.offer_autopilot_settings)?;
+                headers.offer_autopilot_settings = Some(autopilot_json);
+            }
+            _ => {}
+        }
+
+        Ok(headers)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use azure_data_cosmos_driver::models::CosmosRequestHeaders;
+
+    #[test]
+    fn default_throughput_produces_empty_headers() {
+        let tp = ThroughputProperties::default();
+        let headers = CosmosRequestHeaders::try_from(&tp).unwrap();
+        assert!(headers.offer_throughput.is_none());
+        assert!(headers.offer_autopilot_settings.is_none());
+    }
+
+    #[test]
+    fn manual_throughput_sets_offer_throughput() {
+        let tp = ThroughputProperties::manual(400);
+        let headers = CosmosRequestHeaders::try_from(&tp).unwrap();
+        assert_eq!(headers.offer_throughput, Some(400));
+        assert!(headers.offer_autopilot_settings.is_none());
+    }
+
+    #[test]
+    fn autoscale_throughput_sets_autopilot_settings() {
+        let tp = ThroughputProperties::autoscale(4000, None);
+        let headers = CosmosRequestHeaders::try_from(&tp).unwrap();
+        assert!(headers.offer_throughput.is_none());
+        let json = headers
+            .offer_autopilot_settings
+            .expect("should have autopilot settings");
+        assert!(json.contains("4000"));
+    }
+}
